Add files via upload
The Inventory Management In PHP is a simple project developed using PHP, JavaScript, and CSS. The project contains the admin and staff sides. The admin side manages all the management like editing site contents, adding products, staff, customers, etc. The admin/member has to log in to use this system. You can perform the CRUD operations in this project. This project makes a convenient way for the store/inventory admin to keep the records of products and orders in a proper way.