Add files via upload
The Admin Dashboard In PHP is a simple project developed using PHP, JavaScript, and CSS. The project contains only an admin side. The admin does all the work like adding products, amount, editing the details, or deleting the product, vendor, and categories. The admin has to login first. This project makes a convenient way for the store/inventory admin to keep the records of products in a proper way.

About system
This Admin Dashboard is in PHP, JavaScript, Bootstrap, and CSS. Talking about the features of this system, it contains the admin section. All the editings, updating, managing products are from the admin section. This project follows the Crud operation. The design of this system is simple so that the user won’t get any difficulties while working on it.